third_party/blink/web_tests/external/wpt/css/css-values/tree-counting/sibling-function-invalidation-expected.txt

This is a testharness.js-based test.
[FAIL] ::before sibling-count() and sibling-index() after removal
  assert_equals: sibling-index() is 1 after removal expected "10px" but got "20px"
[FAIL] ::after sibling-count() and sibling-index() after removal
  assert_equals: sibling-index() is 2 after removal expected "20px" but got "30px"
[FAIL] ::first-letter sibling-count() and sibling-index() after removal
  assert_equals: sibling-index() is 1 after removal expected "10px" but got "20px"
[FAIL] ::first-line sibling-count() and sibling-index() after removal
  assert_equals: sibling-index() is 2 after removal expected "20px" but got "30px"
Harness: the test ran to completion.
